webnn: change id_to_operand_map to operands array
Using a map is unnecessary here because the ids are just
auto-incrementing integers.

During serialization to the mojo graph, we perform constant folding to
get rid of operands. With the operands array, old operands stay in the
array, they're not referenced by any operations.

Bug: 415953686
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6514349
Commit-Queue: Phillis Tang <phillis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1460377}

+++ services/webnn/public/cpp/webnn_types.h
@@ -11,10 +11,12 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 namespace webnn {
 
-// Key in `GraphInfo.id_to_operand_map`.
-using OperandId = uint64_t;
+// Index into `GraphInfo.operands`.
+// Use uint32_t here because Mojo requires an explicitly sized type and
+// this is big enough.
+using OperandId = uint32_t;
 
-// Index to `GraphInfo.operations`.
+// Index into `GraphInfo.operations`.
 using OperationId = size_t;
 
 }  // namespace webnn
